refactor: bonus function
