+++ simaple/gear/blueprint/gearset_blueprint.py
@@ -6,7 +6,6 @@ from simaple.core import Stat
 from simaple.gear.arcane_symbol import ArcaneSymbol
 from simaple.gear.authentic_symbol import AuthenticSymbol
 from simaple.gear.blueprint.gear_blueprint import PracticalGearBlueprint
-from simaple.gear.gear_repository import GearRepository
 from simaple.gear.gearset import Gearset
 from simaple.gear.potential import PotentialTier
 from simaple.gear.setitem import SetItemRepository
@@ -18,9 +17,7 @@ from simaple.spec.loadable import (  # pylint:disable=unused-import
 
 class GearsetBlueprint(BaseModel, metaclass=TaggedNamespacedABCMeta(kind="blueprint")):
     @abstractmethod
-    def build(
-        self, gear_repository: GearRepository, set_item_repository: SetItemRepository
-    ) -> Gearset:
+    def build(self, set_item_repository: SetItemRepository) -> Gearset:
         ...
